+++ docs/README.md
@@ -1,4 +1,4 @@
-# Sanbot User Guide
+# sanfobot.Sanbot User Guide
 
 // Update the title above to match the actual product name
