+++ src/main/java/sanfobot/command/AddEventTaskCommand.java
@@ -1,3 +1,9 @@
+package sanfobot.command;
+
+import sanfobot.task.EventTask;
+import sanfobot.task.Task;
+import sanfobot.ui.Ui;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
