+++ src/main/java/sanfobot/command/Command.java
@@ -1,3 +1,7 @@
+package sanfobot.command;
+
+import sanfobot.task.Task;
+
 import java.util.ArrayList;
 
 public interface Command {
