+++ src/main/java/sanfobot/command/MarkNotDoneCommand.java
@@ -1,17 +1,21 @@
+package sanfobot.command;
+
+import sanfobot.task.Task;
+import sanfobot.ui.Ui;
+
 import java.util.ArrayList;
 
-public class DeleteTaskCommand {
+public class MarkNotDoneCommand {
     public void execute(String input, ArrayList<Task> taskList) {
         String[] processed = input.split(" ");
         try {
             int num = Integer.parseInt(processed[1].trim()) - 1;
-            Task removed = taskList.get(num);
-            taskList.remove(num);
-            Ui.delete(removed, taskList.size());
+            taskList.get(num).markAsNotDone();
+            Ui.markNotDone(taskList.get(num));
         } catch (ArrayIndexOutOfBoundsException e) {
-            Ui.deleteIndexError();
+            Ui.markNotDoneIndexError();
         } catch (IndexOutOfBoundsException e) {
-            Ui.deleteInvalidIndex();
+            Ui.markNotDoneInvalidIndex();
         }
     }
 }
