+++ src/main/java/sanfobot/task/Task.java
@@ -1,3 +1,5 @@
+package sanfobot.task;
+
 public class Task {
     private String name;
     private boolean isDone;
